Improve StateStore documentation and naming.
